Home

AML

AML web development package in Julia

It automatically creates/extracts HTML/XML files from Julia types!

Installation

Add the package

]add https://github.com/aminya/AML

Usage

Use the package:

using AML

Main macro and I/O

EzXML.parsehtml — Function.

parsehtml(htmlstring)

Parse htmlstring and create an HTML document.

source

EzXML.parsexml — Function.

parsexml(xmlstring)

Parse xmlstring and create an XML document.

source

EzXML.readhtml — Function.

readhtml(filename)

Read filename and create an HTML document.

readhtml(input::IO)

Read input and create an HTML document.

source

EzXML.readxml — Function.

readxml(filename)

Read filename and create an XML document.

readxml(input::IO)

Read input and create an XML document.

source

AML.@aml — Macro.

@aml typedef

Use @aml macro to define a Julia type, and then the package automatically creates a xml or html associated with the defined type.

Type defnition

Use xd"" or hd"" to define a XML or HTML document:

@aml struct Doc xd""

Specify the element name in a string after the struct name

@aml struct Person "person"

Sepecify the html/xml names for childs in strings in front of the struct fields after ,

age::UInt, "age"

For already @aml defined types, name should be the same as its real name

university::University, "university"

If the value is going to be an attribute put a before its name

ID::Int64, a"id"

You can specify the default value for an argument by using = defVal syntax

GPA::Float64 = 4.5, "GPA"

Example 1 - constructor

using AML

@aml struct Person "person"
    age::UInt, "age"
    field::String, "study-field"
    GPA::Float64 = 4.5, "GPA"
    courses::Vector{String}, "taken-courses"
    ID::Int64, a"id"
end

@aml struct University "university"
    name, a"university-name"
    people::Vector{Person}, "person"
end

@aml struct Doc xd""
    university::University, "university"
end

P1 = Person(age=24, field="Mechanical Engineering", courses=["Artificial Intelligence", "Robotics"], ID = 1)
P2 = Person(age=18, field="Computer Engineering", GPA=4, courses=["Julia"], ID = 2)

U = University(name="Julia University", people=[P1, P2])

D = Doc(university = U)

julia> print(P1.aml)
<person id="1">
  <age>24</age>
  <study-field>Mechanical Engineering</study-field>
  <GPA>4.5</GPA>
  <taken-courses>Artificial Intelligence</taken-courses>
  <taken-courses>Robotics</taken-courses>
</person>

julia> print(P2.aml)
<person id="2">
  <age>18</age>
  <study-field>Computer Engineering</study-field>
  <GPA>4</GPA>
  <taken-courses>Julia</taken-courses>
</person>

julia> print(U.aml)
<university university-name="Julia University">
  <person id="1">
    <age>24</age>
    <study-field>Mechanical Engineering</study-field>
    <GPA>4.5</GPA>
    <taken-courses>Artificial Intelligence</taken-courses>
    <taken-courses>Robotics</taken-courses>
  </person>
  <person id="2">
    <age>18</age>
    <study-field>Computer Engineering</study-field>
    <GPA>4</GPA>
    <taken-courses>Julia</taken-courses>
  </person>
</university>

julia> print(D.aml)
<?xml version="1.0" encoding="UTF-8" standalone="yes"?><!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN" "http://www.w3.org/TR/REC-html40/loose.dtd">
<university university-name="Julia University">
  <person id="1">
    <age>24</age>
    <study-field>Mechanical Engineering</study-field>
    <GPA>4.5</GPA>
    <taken-courses>Artificial Intelligence</taken-courses>
    <taken-courses>Robotics</taken-courses>
  </person>
  <person id="2">
    <age>18</age>
    <study-field>Computer Engineering</study-field>
    <GPA>4</GPA>
    <taken-courses>Julia</taken-courses>
  </person>
</university>

Example 2 - extractor

using AML

xml = parsexml("""
<?xml version="1.0" encoding="UTF-8" standalone="yes"?><!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN" "http://www.w3.org/TR/REC-html40/loose.dtd">
<university university-name="Julia University">
  <person id="1">
    <age>24</age>
    <study-field>Mechanical Engineering</study-field>
    <GPA>4.5</GPA>
    <taken-courses>Artificial Intelligence</taken-courses>
    <taken-courses>Robotics</taken-courses>
  </person>
  <person id="2">
    <age>18</age>
    <study-field>Computer Engineering</study-field>
    <GPA>4</GPA>
    <taken-courses>Julia</taken-courses>
  </person>
</university>
""")

@aml struct Person "person"
    age::UInt, "age"
    field::String, "study-field"
    GPA::Float64 = 4.5, "GPA"
    courses::Vector{String}, "taken-courses"
    ID::Int64, a"id"
end

@aml struct University "university"
    name, a"university-name"
    people::Vector{Person}, "person"
end

@aml struct Doc xd""
    university::University, "university"
end

# extract Doc

D = Doc(xml)

# extract University

U = University(D.university)

julia>U.name
"Julia University"

# extract Person

P1 = Person(U.people[1])

julia>P1.age
24

julia>P1.field
Mechanical Engineering

julia>P1.GPA
4.5

julia>P1.courses
["Artificial Intelligence", "Robotics"]

julia>P1.ID
1

P2 = Person(U.people[2])

source

AML.findallcontent

AML.findalllocal

AML.findfirstcontent

AML.findfirstlocal

EzXML.parsehtml

EzXML.parsexml

EzXML.readhtml

EzXML.readxml

AML.@aml

Backend utilities

AML.findallcontent — Method.

findallcontent(type, string, node)

Finds all the elements with the address of string in the node, and converts the elements to Type object.

source

AML.findalllocal — Method.

findalllocal(s,node)

findalllocal with ignoring namespaces. It considers element.name for returning the elements

source

AML.findfirstcontent — Method.

findfirstcontent(element,node)
findfirstcontent(type,element,node)

Returns first element content. It also convert to the desired format by passing type. element is given as string.

findfirstcontent("/instrument-name",node)
findfirstcontent(UInt8,"/midi-channel",node)

source

AML.findfirstlocal — Method.

findfirstlocal(s, node)

findfirst with ignoring namespaces. It considers element.name for returning the elements

source