Syntax Reference

AcuteML.newTemplate

AcuteML.render2file

AcuteML.@aml

Main Macro and IO

AcuteML.@aml — Macro.

@aml typedef

Type Definition

Use @aml macro to define a Julia type, and then the package automatically creates a xml or html associated with the defined type.

Document Definition

Use xd"" or hd"" to define a XML or HTML document:

  @aml mutable struct Doc xd""
    # add fields (elements) here
  end
  ```

  ### Nodes (Elements) Definition
  * Specify the html/xml struct name as a string after the struct name after a space
  ```julia
  @aml mutable struct Person "person"
    # add fields (elements) here
  end
  ```
  * If the html/xml name is the same as struct name, you can use `"~"` instead
  ```julia
  @aml mutable struct person "~"
    # add fields (elements) here
  end
  ```

  ### Fields Names
  * Sepecify the html/xml field name as a string in front of the field after `,`
  ```julia
  field, "study-field"
  ```
  * If the html/xml name is the same as variable name, you can use `"~"` instead
  ```julia
  age::UInt, "~"
  ```

  ### Attributes
  * If the value is going to be an attribute put `a` before its name
  ```julia
  id::Int64, a"~"
  ```

  ### Default Value
  * You can specify the default value for an argument by using `= defVal` syntax
  ```julia
  GPA::Float64 = 4.5, "~"
  ```

  ### Value Types
  You can use Julia types or  defined types for values.

  * If you don't specify the type of a variable, it is considered to be string for aml manipulations:
  ```julia
  field, "study-field"
  ```
  However, for a high performance code specify String type (`field::String, "study-field"`)

  * For already `@aml` defined types, name should be the same as the defined type root name
  ```julia
  university::University, "university"
  ```

  ### Value Checking
  You can define any restriction for values using functions.

  * To define any restrictions for the values of one field, define a function that checks a criteria for the field value and returns Bool, and put its name after a `,` after the field name:
  ```julia
  GPA::Float64, "~", GPAcheck
  ```

  * To define any restrictions for multiple values of a struct, define a function that gets all the variables and checks a criteria and returns Bool, and put its name after a `,` after the struct name:
  ```julia
  @aml mutable struct Person "person", courseCheck
  # ...
  end
  ```

  Refer to https://aminya.github.io/AcuteML.jl/dev/valueChecking/ for some of these functions examples.

  ### Optional Fields
  * If a field is optional, don't forget to define its type as `UN{}` (Union with Nothing), and set the default value as `nothing`.
  ```julia
  residence::UN{String}=nothing, "residence-stay" # optional with nothing as default value
  ```
  ```julia
  funds::UN{String}, "financial-funds"   # optional, but you should pass nothing manually in construction
  ```

  ### Empty Elements (Self-Closing) Definition
  * Use `sc"name"` to define a self-closing (empty) element (e.g. `<rest />`)
  ```julia
  @aml struct rest sc"~"
  end
  ```
  -------------------------------------------------------

  # Example - Struct Definition

  First, we define the structs using `@aml` to store the data in:

  ```julia
  using AcuteML

  # Types definition

  # Person Type
  @aml mutable struct Person "person", courseCheck
      age::UInt64, "~"
      field, "study-field"
      GPA::Float64 = 4.5, "~", GPAcheck
      courses::Vector{String}, "taken-courses"
      id::Int64, a"~"
  end

  @aml mutable struct University "university"
      name, a"university-name"
      people::Vector{Person}, "person"
  end

  @aml mutable struct Doc xd""
      university::University, "~"
  end

  ```

  ```julia
  # Value Checking Functions
  GPAcheck(x) = x <= 4.5 && x >= 0

  function courseCheck(age, field, GPA, courses, id)

      if field == "Mechanical Engineering"
          relevant = ["Artificial Intelligence", "Robotics", "Machine Design"]
      elseif field == "Computer Engineering"
          relevant = ["Julia", "Algorithms"]
      else
          error("study field is not known")
      end

      return any(in.(courses, Ref(relevant)))
  end
  ```
  -------------------------------------------------------

  # Example - Constructor

  After we defined the structs, we can create instances of them by passing our data to the fields:

  ```julia

  P1 = Person(age=24, field="Mechanical Engineering", courses=["Artificial Intelligence", "Robotics"], id = 1)
  P2 = Person(age=18, field="Computer Engineering", GPA=4, courses=["Julia"], id = 2)

  U = University(name="Julia University", people=[P1, P2])

  D = Doc(university = U)

  D.university.people[2].GPA=4.2 # mutability support after Doc creation

  ```

  ```julia
  # An example that doesn't meet the criteria function for GPA because GPA is more than 4.5
  P3 = Person(age=99, field="Macro Wizard", GPA=10, courses=["Julia Magic"], id = 3)
  julia>
  GPA doesn't meet criteria function
  ```

  ```html
  julia> pprint(P1) # or print(P1.aml)
  <person id="1">
    <age>24</age>
    <study-field>Mechanical Engineering</study-field>
    <GPA>4.5</GPA>
    <taken-courses>Artificial Intelligence</taken-courses>
    <taken-courses>Robotics</taken-courses>
  </person>

  julia> pprint(U) # or print(U.aml)
  <university university-name="Julia University">
    <person id="1">
      <age>24</age>
      <study-field>Mechanical Engineering</study-field>
      <GPA>4.5</GPA>
      <taken-courses>Artificial Intelligence</taken-courses>
      <taken-courses>Robotics</taken-courses>
    </person>
    <person id="2">
      <age>18</age>
      <study-field>Computer Engineering</study-field>
      <GPA>4.2</GPA>
      <taken-courses>Julia</taken-courses>
    </person>
  </university>

  julia> pprint(D) # or print(D.aml)
  <?xml version="1.0" encoding="UTF-8" standalone="yes"?><!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN" "http://www.w3.org/TR/REC-html40/loose.dtd">
  <university university-name="Julia University">
    <person id="1">
      <age>24</age>
      <study-field>Mechanical Engineering</study-field>
      <GPA>4.5</GPA>
      <taken-courses>Artificial Intelligence</taken-courses>
      <taken-courses>Robotics</taken-courses>
    </person>
    <person id="2">
      <age>18</age>
      <study-field>Computer Engineering</study-field>
      <GPA>4.2</GPA>
      <taken-courses>Julia</taken-courses>
    </person>
  </university>
  ```
-------------------------------------------------------

# Example - Extractor

After we defined the structs, we can automatically extract and store the data in their fields:

julia using AcuteML

xml = parsexml(""" <?xml version="1.0" encoding="UTF-8" standalone="yes"?><!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN" "http://www.w3.org/TR/REC-html40/loose.dtd"> <university university-name="Julia University"> <person id="1"> <age>24</age> <study-field>Mechanical Engineering</study-field> <GPA>4.5</GPA> <taken-courses>Artificial Intelligence</taken-courses> <taken-courses>Robotics</taken-courses> </person> <person id="2"> <age>18</age> <study-field>Computer Engineering</study-field> <GPA>4</GPA> <taken-courses>Julia</taken-courses> </person> </university> """)

extract Doc

D = Doc(xml) # StructName(xml) like Doc(xml) extracts the data and stores them in proper format

Now you can access all of the data by calling the fieldnames

extract University

U = D.university

julia>U.name "Julia University"

extract Person

P1 = U.people[1]

julia>P1.age 24

julia>P1.field Mechanical Engineering

julia>P1.GPA 4.5

julia>P1.courses ["Artificial Intelligence", "Robotics"]

julia>P1.id 1 ```

source

Templating

AcuteML.newTemplate — Function.

newTemplate(name)

Create new destination html file as the template

newTemplate(name, :function)

Prints a function to be used as a template

Examples

# you can create a file and edit the file directly by using
newTemplate("person")

## create person function to store out html template
newTemplate("person", :function)

source

AcuteML.render2file — Function.

render2file(destination, overwrite, var...)

render variables passed as an input to the destination file.

You should put var in the destination file/string so var is evaluated there. Pass the variables as keyword arguments with the same name you used in the html string/file. Variables should be string,

If you want to statically overwrite the file pass true as the 2nd argument to the function. Useful if you don't want a dynamic website.

Examples

# Add the following html code to the generated html file
#=
<person id=$(id)>
  <age>$(age)</age>
  <study-field>$(field)</study-field>
  <GPA>$(GPA)</GPA>
  <taken-courses>$(courses[1])</taken-courses>
  <taken-courses>$(courses[2])</taken-courses>
</person>
=#

# Specify the template (or its path), and also the variables for rendering
out =render2file("person", false,
  id = 1,
  age = 24,
  field = "Mechanical Engineering",
  GPA = 4.5,
  courses = ["Artificial Intelligence", "Robotics"])

# you pass `true` as the 2nd argument to owerwrite person.html statically.

source

Backend utilities