Syntax Reference

AcuteML.addelementOne!

AcuteML.addelementVect!

AcuteML.docOrElmInit

AcuteML.findallcontent

AcuteML.findalllocal

AcuteML.findfirstcontent

AcuteML.findfirstlocal

AcuteML.newTemplate

AcuteML.render2file

AcuteML.updateallcontent!

AcuteML.updatefirstcontent!

EzXML.parsehtml

EzXML.parsexml

EzXML.readhtml

EzXML.readxml

AcuteML.@aml

Main Macro and IO

EzXML.parsehtml — Function.

parsehtml(htmlstring)

Parse htmlstring and create an HTML document.

source

EzXML.parsexml — Function.

parsexml(xmlstring)

Parse xmlstring and create an XML document.

source

EzXML.readhtml — Function.

readhtml(filename)

Read filename and create an HTML document.

readhtml(input::IO)

Read input and create an HTML document.

source

EzXML.readxml — Function.

readxml(filename)

Read filename and create an XML document.

readxml(input::IO)

Read input and create an XML document.

source

AcuteML.@aml — Macro.

@aml typedef

Type defnition

Use @aml macro to define a Julia type, and then the package automatically creates a xml or html associated with the defined type.

Document Defnition

Use xd"" or hd"" to define a XML or HTML document:

@aml mutable struct Doc xd""
# add fields (elements) here
end

Nodes (Elements) Defnition

Specify the html/xml struct name as a string after the struct name after a space

@aml mutable struct Person "person"
# add fields (elements) here
end

If the html/xml name is the same as struct name, you can use "~" instead

@aml mutable struct person "~"
# add fields (elements) here
end

Fields Names

Sepecify the html/xml field name as a string in front of the field after ,

field, "study-field"

If the html/xml name is the same as variable name, you can use "~" instead

age::UInt, "~"

Attributes

If the value is going to be an attribute put a before its name

id::Int64, a"~"

Default Value

You can specify the default value for an argument by using = defVal syntax

GPA::Float64 = 4.5, "~"

Value Types

You can use Julia types or defined types for values.

If you don't specify the type of a variable, it is considered to be string:

field, "study-field"

For already @aml defined types, name should be the same as the defined type root name

university::University, "university"

Value Checking

You can define any restriction for values using functions.

To define any restrictions for the values of one field, define a function that checks a criteria for the field value and returns Bool, and put its name after a , after the field name:

GPA::Float64, "~", GPAcheck

To define any restrictions for multiple values of a struct, define a function that gets all the variables and checks a criteria and returns Bool, and put its name after a , after the struct name:

@aml mutable struct Person "person", courseCheck
# ...
end

Refer to https://aminya.github.io/AcuteML.jl/dev/valueChecking/ for some of these functions examples.

Optional Fields

If a field is optional, don't forget to define its type as UN{} (Union with Nothing), and set the default value as nothing.

residence::UN{String}=nothing, "residence-stay" # optional with nothing as default value

funds::UN{String}, "financial-funds"   # optional, but you should pass nothing manually in construction

Empty Elements (Self-Closing) Defnition

Use sc"name" to define a self-closing (empty) element (e.g. <rest />)

@aml struct rest sc"~"
end

Example - Type Definition

using AcuteML

# Types definition

# Person Type
@aml mutable struct Person "person", courseCheck
    age::UInt64, "~"
    field, "study-field"
    GPA::Float64 = 4.5, "~", GPAcheck
    courses::Vector{String}, "taken-courses"
    id::Int64, a"~"
end

@aml mutable struct University "university"
    name, a"university-name"
    people::Vector{Person}, "person"
end

@aml mutable struct Doc xd""
    university::University, "~"
end

# Value Checking Functions
GPAcheck(x) = x <= 4.5 && x >= 0

function courseCheck(age, field, GPA, courses, id)

    if field == "Mechanical Engineering"
        relevant = ["Artificial Intelligence", "Robotics", "Machine Design"]
    elseif field == "Computer Engineering"
        relevant = ["Julia", "Algorithms"]
    else
        error("study field is not known")
    end

    return any(in.(courses, Ref(relevant)))
end

Example - Constructor

P1 = Person(age=24, field="Mechanical Engineering", courses=["Artificial Intelligence", "Robotics"], id = 1)
P2 = Person(age=18, field="Computer Engineering", GPA=4, courses=["Julia"], id = 2)

P2.GPA=4.2 # mutability support

U = University(name="Julia University", people=[P1, P2])

D = Doc(university = U)

# An example that doesn't meet the critertia function for GPA because GPA is more than 4.5
P3 = Person(age=99, field="Macro Wizard", GPA=10, courses=["Julia Magic"], id = 3)
julia>
GPA doesn't meet criteria function

julia> print(P1.aml)
<person id="1">
  <age>24</age>
  <study-field>Mechanical Engineering</study-field>
  <GPA>4.5</GPA>
  <taken-courses>Artificial Intelligence</taken-courses>
  <taken-courses>Robotics</taken-courses>
</person>

julia> print(U.aml)
<university university-name="Julia University">
  <person id="1">
    <age>24</age>
    <study-field>Mechanical Engineering</study-field>
    <GPA>4.5</GPA>
    <taken-courses>Artificial Intelligence</taken-courses>
    <taken-courses>Robotics</taken-courses>
  </person>
  <person id="2">
    <age>18</age>
    <study-field>Computer Engineering</study-field>
    <GPA>4.2</GPA>
    <taken-courses>Julia</taken-courses>
  </person>
</university>

julia> print(D.aml)
<?xml version="1.0" encoding="UTF-8" standalone="yes"?><!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN" "http://www.w3.org/TR/REC-html40/loose.dtd">
<university university-name="Julia University">
  <person id="1">
    <age>24</age>
    <study-field>Mechanical Engineering</study-field>
    <GPA>4.5</GPA>
    <taken-courses>Artificial Intelligence</taken-courses>
    <taken-courses>Robotics</taken-courses>
  </person>
  <person id="2">
    <age>18</age>
    <study-field>Computer Engineering</study-field>
    <GPA>4.2</GPA>
    <taken-courses>Julia</taken-courses>
  </person>
</university>

Example - Extractor

using AcuteML

xml = parsexml("""
<?xml version="1.0" encoding="UTF-8" standalone="yes"?><!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN" "http://www.w3.org/TR/REC-html40/loose.dtd">
<university university-name="Julia University">
  <person id="1">
    <age>24</age>
    <study-field>Mechanical Engineering</study-field>
    <GPA>4.5</GPA>
    <taken-courses>Artificial Intelligence</taken-courses>
    <taken-courses>Robotics</taken-courses>
  </person>
  <person id="2">
    <age>18</age>
    <study-field>Computer Engineering</study-field>
    <GPA>4</GPA>
    <taken-courses>Julia</taken-courses>
  </person>
</university>
""")

# extract Doc
D = Doc(xml) # StructName(xml) like Doc(xml) extracts the data and stores them in proper format

# Now you can access all of the data by calling the fieldnames

# extract University
U = D.university

julia>U.name
"Julia University"

# extract Person

P1 = U.people[1]

julia>P1.age
24

julia>P1.field
Mechanical Engineering

julia>P1.GPA
4.5

julia>P1.courses
["Artificial Intelligence", "Robotics"]

julia>P1.id
1

source

Templating

AcuteML.newTemplate — Function.

newTemplate(name)

Create new destination html file as the template

newTemplate(name, :function)

Prints a function to be used as a template

Examples

# you can create a file and edit the file directly by using
newTemplate("person")

## create person function to store out html template
newTemplate("person", :function)

source

AcuteML.render2file — Function.

render2file(destination, overwrite, var...)

render variables passed as an input to the destination file.

You should put var in the destination file/string so var is evaluated there. Pass the variables as keyword arguments with the same name you used in the html string/file. Variables should be string,

If you want to statically overwrite the file pass true as the 2nd argument to the function. Useful if you don't want a dynamic website.

Examples

# Add the following html code to the generated html file
#=
<person id=$(id)>
  <age>$(age)</age>
  <study-field>$(field)</study-field>
  <GPA>$(GPA)</GPA>
  <taken-courses>$(courses[1])</taken-courses>
  <taken-courses>$(courses[2])</taken-courses>
</person>
=#

# Specify the template (or its path), and also the variables for rendering
out =render2file("person", false,
  id = 1,
  age = 24,
  field = "Mechanical Engineering",
  GPA = 4.5,
  courses = ["Artificial Intelligence", "Robotics"])

# you pass `true` as the 2nd argument to owerwrite person.html statically.

source

Backend utilities

AcuteML.addelementOne! — Method.

addelementOne!(node, name, value, amlType)

Add one element to a node/document ```

source

AcuteML.addelementVect! — Method.

addelementVect!(node, name, value, amlType)

Add a vector to a node/document ```

source

AcuteML.docOrElmInit — Function.

docOrElmInit(name)
docOrElmInit(type, name)

Function to initialize the aml

type: 0 : element node # default 10: empty element node -1: xml -2: html

source

AcuteML.findallcontent — Method.

findallcontent(type, string, node, amlType)

Finds all the elements with the address of string in the node, and converts the elements to Type object.

findallcontent(UInt8,"/midi-channel",node, 0)

source

AcuteML.findalllocal — Method.

findfirstlocal(string, node)

findalllocal with ignoring namespaces. It considers element.name for returning the elements

source

AcuteML.findfirstcontent — Method.

findfirstcontent(element,node, amlType)
findfirstcontent(type,element,node, amlType)

Returns first element content. It also convert to the desired format by passing type. element is given as string.

findfirstcontent("/instrument-name",node, 0)
findfirstcontent(UInt8,"/midi-channel",node, 0)

source

AcuteML.findfirstlocal — Method.

findfirstlocal(string, node)

findfirst with ignoring namespaces. It considers element.name for returning the elements

source

AcuteML.updateallcontent! — Method.

updateallcontent!(value, string, node, amlType)

Finds all the elements with the address of string in the node, and updates the content

source

AcuteML.updatefirstcontent! — Method.

updatefirstcontent(value, string, node, amlType)

Updates first element content. It also converts any type to string. element is given as string.

source